requirements.txt
altgraph==0.17.4
GPUtil==1.4.0
packaging==24.1
pillow==10.4.0
psutil==6.0.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
ttkthemes==3.2.2
